+++ tests/test_files.py
@@ -7,6 +7,7 @@ from incydr import Client
 
 
 TEST_SHA256 = "38acb15d02d5ac0f2a2789602e9df950c380d2799b4bdb59394e4eeabdd3a662"
+TEST_XFC_ID = "test"
 BAD_SHA256 = "asdf"
 TEST_DATA = b"test data"
 
@@ -15,6 +16,14 @@ TEST_DATA = b"test data"
 def mock_file_download(httpserver_auth: HTTPServer):
     httpserver_auth.expect_request(
         f"/v1/files/get-file-by-sha256/{TEST_SHA256}"
+    ).respond_with_data(response_data=TEST_DATA, status=200)
+    return httpserver_auth
+
+
+@pytest.fixture
+def mock_file_download_by_xfc(httpserver_auth: HTTPServer):
+    httpserver_auth.expect_request(
+        f"/v1/files/get-file-by-xfc-content-id/{TEST_XFC_ID}"
     ).respond_with_data(response_data=TEST_DATA, status=200)
     return httpserver_auth
 
@@ -31,6 +40,14 @@ def mock_bad_sha256(httpserver_auth: HTTPServer):
 def mock_file_not_found(httpserver_auth: HTTPServer):
     httpserver_auth.expect_request(
         f"/v1/files/get-file-by-sha256/{TEST_SHA256}"
+    ).respond_with_data(response_data="", status=404)
+    return httpserver_auth
+
+
+@pytest.fixture
+def mock_file_not_found_by_xfc_id(httpserver_auth: HTTPServer):
+    httpserver_auth.expect_request(
+        f"/v1/files/get-file-by-xfc-content-id/{TEST_XFC_ID}"
     ).respond_with_data(response_data="", status=404)
     return httpserver_auth
 
@@ -66,6 +83,35 @@ def test_download_file_by_sha256_raises_error_when_file_not_found(mock_file_not_
 def test_stream_file_by_sha256_streams_file(mock_file_download):
     c = Client()
     response = c.files.v1.stream_file_by_sha256(TEST_SHA256)
+    content = b"".join([c for c in response.iter_content(chunk_size=128) if c])
+    assert content == TEST_DATA
+
+
+def test_download_file_by_xfc_id_calls_with_correct_parameter(
+    mock_file_download_by_xfc, tmp_path
+):
+    c = Client()
+    p = tmp_path / "testfile.test"
+    f = c.files.v1.download_file_by_xfc_content_id(TEST_XFC_ID, p)
+    with open(f, "rb") as file:
+        content = file.read()
+    assert content == TEST_DATA
+    mock_file_download_by_xfc.check()
+
+
+def test_download_file_by_xfc_id_raises_error_when_file_not_found(
+    mock_file_not_found_by_xfc_id,
+):
+    c = Client()
+    with pytest.raises(HTTPError) as error:
+        c.files.v1.download_file_by_xfc_content_id(TEST_XFC_ID, "testpath.text")
+    assert error.value.response.status_code == 404
+    mock_file_not_found_by_xfc_id.check()
+
+
+def test_stream_file_by_xfc_id_streams_file(mock_file_download_by_xfc):
+    c = Client()
+    response = c.files.v1.stream_file_by_xfc_content_id(TEST_XFC_ID)
     content = b"".join([c for c in response.iter_content(chunk_size=128) if c])
     assert content == TEST_DATA